Created multiple templates and added logout
