+++ src/main/java/com/seifernet/wissen/controller/WissenController.java
@@ -3,12 +3,10 @@ package com.seifernet.wissen.controller;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.seifernet.wissen.util.ModelAttributes;
 import com.seifernet.wissen.util.URL;
 import com.seifernet.wissen.util.WebResources;
-import com.seifernet.wissen.rpg.Dice;
 
 /**
  * Application controller
@@ -25,42 +23,13 @@ public class WissenController {
 	 * @return
 	 */
 	@RequestMapping(URL.INDEX)
-	private String index( Model model ) {
-		model.addAttribute(ModelAttributes.CONTENT_SOURCE, WebResources.CONTENT);
-		model.addAttribute(ModelAttributes.CONTENT_FRAGMENT, "index");
+	private String index(Model model) {
+		model.addAttribute(ModelAttributes.CONTENT_SOURCE, WebResources.INDEX);
+		model.addAttribute(ModelAttributes.CONTENT_FRAGMENT, "content");
 		
 		model.addAttribute(ModelAttributes.CUSTOM_JS_SOURCE, WebResources.CUSTOM_JS);
 		model.addAttribute(ModelAttributes.CUSTOM_JS_FRAGMENT, "indexjs");
 		
 		return WebResources.BASE_LAYOUT;
 	}
-	
-	@ResponseBody
-	@RequestMapping("/integrations/dice")
-	private Integer dice(Model model){
-		return Dice.rollDice(6);
-	}
-	
-	@RequestMapping(URL.SYS_DASHBOARD)
-	private String sysDashboard( Model model ) {
-		model.addAttribute(ModelAttributes.CONTENT_SOURCE, WebResources.CONTENT);
-		model.addAttribute(ModelAttributes.CONTENT_FRAGMENT, "sysdashboard");
-		
-		model.addAttribute(ModelAttributes.CUSTOM_JS_SOURCE, WebResources.CUSTOM_JS);
-		model.addAttribute(ModelAttributes.CUSTOM_JS_FRAGMENT, "empty");
-		
-		return WebResources.BASE_LAYOUT;
-	}
-
-	@RequestMapping(URL.KNOWLEDGE)
-	private String knowledge( Model model ) {
-		model.addAttribute(ModelAttributes.CONTENT_SOURCE, WebResources.CONTENT);
-		model.addAttribute(ModelAttributes.CONTENT_FRAGMENT, "knowledge");
-		
-		model.addAttribute(ModelAttributes.CUSTOM_JS_SOURCE, WebResources.CUSTOM_JS);
-		model.addAttribute(ModelAttributes.CUSTOM_JS_FRAGMENT, "knowledgejs");
-		
-		return WebResources.BASE_LAYOUT;
-	}
-	
 }
