+++ src/main/java/com/seifernet/wissen/util/URL.java
@@ -8,6 +8,4 @@ package com.seifernet.wissen.util;
 public class URL {
 	
 	public static final String INDEX = "/";
-	public static final String KNOWLEDGE = "/knowledge";
-	public static final String SYS_DASHBOARD = "/sysdashboard";
 }
