+++ src/main/java/com/seifernet/wissen/util/WebResources.java
@@ -9,6 +9,6 @@ public class WebResources {
 	
 	public static final String BASE_LAYOUT 	= "main_layout";
 	public static final String VOID			= "empty";
-	public static final String CONTENT 		= "content";
+	public static final String INDEX		= "index";
 	public static final String CUSTOM_JS 	= "custom_js";
 }
