+++ src/main/resources/public/js/wissen.js
@@ -1,113 +1,36 @@
-/**
- * Utility function to format date values from Javascript Date object
- */
-Number.prototype.pad = function(size) {
-	var s = String(this);
-	while (s.length < (size || 2)) {s = "0" + s;}
-	return s;
-};
-
-
-/**
- *  Utility function to format date values to compact 3 digit time difference from current date
- */
-function dateToStringCompact(a,b){
-	var milis = a.getTime() - b.getTime();
-	
-	if(milis > 31536000000) {
-		return (Math.trunc(milis/31536000000)).pad(2) + "y";
-	} else if(milis > 2592000000) {
-		return (Math.trunc(milis/2592000000)).pad(2) + "M";
-	} else if(milis > 86400000) {
-		return (Math.trunc(milis/86400000)).pad(2) + "d";
-	} else if(milis > 3600000){
-		return (Math.trunc(milis/3600000)).pad(2) + "h";
-	} else if(milis > 60000){
-		return (Math.trunc(milis/60000)).pad(2) + "m";
+/** Load of general page elements **/
+function refreshGenericPageElements(validToken){
+	if(validToken){
+		$("#authuser").text($.cookie("authuser"));
+		
+		$("#authuseritem").attr('onclick','').unbind('click');
+		var usermenu = $("<div></div>");
+		$("#authuseritem").children("div").remove();
+		usermenu.addClass("menu");
+		$("#authuseritem").append(usermenu);
+		var logoutItem = $("<div></div>");
+		logoutItem.addClass("item");
+		logoutItem.append($("<span onclick='logout()'>Logout</span>"));
+		usermenu.append(logoutItem);
+		$("#authuseritem").addClass("dropdown");
+		$('.ui.dropdown').dropdown();
+		
 	} else {
-		return "ovd";
+		$("#authuseritem").attr('onclick','').unbind('click');
+		var bck = $('<div id="authuseritem"></div>');
+		bck.addClass('ui right item');
+		bck.click(function(){showAuthModal("menuButton")});
+		bck.append('<i class="large user circle outline icon"></i><span id="authuser">Anonymous</span>');
+		$("#authuseritem").replaceWith(bck);
 	}
-}
-
-/**
- *  Utility function to format date values to compact 3 digit time difference from current date
- */
-function dateToString(a,b){
-	var milis = a.getTime() - b.getTime();
-	
-	if(milis > 31536000000) {
-		if(Math.trunc(milis/31536000000) > 1){
-			return  (Math.trunc(milis/31536000000)) + " years"
-		} else {
-			return "1 year";
-		}
-	} else if(milis > 2592000000) {
-		if(Math.trunc(milis/2592000000) > 1){
-			return  (Math.trunc(milis/2592000000)) + " months"
-		} else {
-			return "1 month";
-		}
-	} else if(milis > 86400000) {
-		if(Math.trunc(milis/86400000) > 1){
-			return  (Math.trunc(milis/86400000)) + " days"
-		} else {
-			return "1 day";
-		}
-	} else if(milis > 3600000){
-		if(Math.trunc(milis/3600000) > 1){
-			return  (Math.trunc(milis/3600000)) + " hours"
-		} else {
-			return "1 hour";
-		}
-	} else if(milis > 60000){
-		if(Math.trunc(milis/60000) > 1){
-			return  (Math.trunc(milis/60000)) + " minutes"
-		} else {
-			return "1 minute";
-		}
-	} else {
-		return "seconds";
-	}
-}
-
-/**
- * Utility function to get label color based on due date
- * 
- * @param fdate
- * @returns
- */
-function getLabelColor(fdate){
-	var actualDate = new Date();
-	var providedDate = new Date(fdate);
-	var milis = providedDate.getTime() - actualDate.getTime();
-	
-	if(milis > 18000000){
-		return "teal";
-	} else if (milis > 7200000){
-		return "yellow";
-	} else {
-		return "red";
-	}
-}
-
-/**
- * Utility function to format date values
- * 
- * @param fdate
- * @returns
- */
-function formatDate(fdate){
-	var providedDate = new Date(fdate);
-	return (providedDate.getMonth()+1).pad(2) + "/"+(providedDate.getDate()).pad(2)+"/"+providedDate.getFullYear() + " " + (providedDate.getHours()).pad(2) + ":"+(providedDate.getMinutes()).pad(2);
 }
 
 /**
  * Session token validation
  */
 //<![CDATA[
-function validateToken(callback){
+function validateToken(){
 	if($.cookie("authuser") != undefined && $.cookie("authtoken") != undefined) {
-		valid = false;
 		$.ajax({
 			type: 'GET',
 			url: "/oauth/check_token?token=" + $.cookie("authtoken"),
@@ -116,42 +39,44 @@ function validateToken(callback){
 				"Accept" : "application/json"
 			},
 			success: function(response) {
-				callback(true);
+				refreshGenericPageElements(true);
+				loadAuthContent();
 			},
 			error: function(XMLHttpRequest) {
-				if(XMLHttpRequest.status === 400) {
-					callback(false);
-				}
 				$.removeCookie("authuser");
 				$.removeCookie("authtoken");
-				//TODO implement other error status handling
+				if(XMLHttpRequest.status === 400) {
+					refreshGenericPageElements(false);
+					loadAnonContent();
+				}
 			}
 		});
 	} else {
-		callback(false);
+		refreshGenericPageElements(false);
+		loadAnonContent();
 	}
 }
 //]]>
 
 /*eslint-disable no-undef */
 $('.ui.dropdown').dropdown();
-$('.ui.modal').modal({
-	allowMultiple: false
-});
+$('.ui.modal').modal({allowMultiple: false});
 $('#regmod').modal('attach events', '#tgrm');
 $('.progress').progress();
-$('.message .close').on('click', function() {
-	$(this).closest('.message').transition('fade');
-});
+$('.message .close').on('click', function(){$(this).closest('.message').transition('fade');});
 $('.ui.accordion').accordion();
-var globalTimeOut = null;
+
+/**Main UI load function**/
+$(document).ready(function(){
+	validateToken();
+});
+
 
 /**
  * Used to raise the modal for authentication
  */
 function showAuthModal(origin){
 	if($.cookie("authuser") === undefined && $.cookie("authtoken") === undefined) {
-		//Modal init
 		$("#wrongcredentials").addClass("hidden");
 		$("#user").val("");
 		$("#passwd").val("");
@@ -164,6 +89,13 @@ function showAuthModal(origin){
 		}
 		$("#authmod").modal("show");
 	}
+}
+
+function logout(){
+	$.removeCookie("authuser");
+	$.removeCookie("authtoken");
+	refreshGenericPageElements(false);
+	loadAnonContent();
 }
 
 /**
@@ -195,7 +127,8 @@ function authenticate(){
 			$.cookie("authtoken", resultData.access_token);
 			$.cookie("authuser", user);
 			$("#authmod").modal("hide");
-			validateToken(refreshPageElements);
+			refreshGenericPageElements(true)
+			loadAuthContent();
 		}
 	});
 }
