+++ src/main/resources/templates/custom_js.html
@@ -4,14 +4,9 @@
     <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
   </head>
   <body>
-    
     <div th:fragment="indexjs">
     <script src="/js/index.js"></script>
     </div>
-    
-    
-    <div th:fragment="knowledgejs">
-    <script src="/js/knowledge.js"></script>
-    </div>
+
   </body>
 </html>
