+++ src/main/resources/templates/index.html
@@ -5,7 +5,9 @@
   </head>
   <body>
   
-    <th:block th:fragment="empty">
+    <th:block th:fragment="content">
+    <div class="ui container">
+    </div>
     </th:block>
 
   </body>
