+++ src/main/resources/templates/page_menu.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+  <head>
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
+  </head>
+  <body>
+  
+    <th:block th:fragment="content">
+    <div class="ui top attached menu">
+      <div class="ui item">
+        <a href="/"><h3 class="ui header">Wissen</h3></a>
+      </div>
+      <div class="ui right item" id="authuseritem" onclick="showAuthModal('menuButton')">
+        <i class="large user circle outline icon"></i> <span id="authuser">Anonymous</span>
+      </div>
+    </div>
+    </th:block>
+    
+  </body>
+</html>
+
+
+
