+++ src/main/resources/templates/register_modal.html
@@ -1,0 +1,44 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+  <head>
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
+  </head>
+  <body>
+  
+    <th:block th:fragment="content">
+    <div class="ui modal" id="regmod">
+      <i class="close icon"></i>
+      <div class="header">Register</div>
+      
+      <div class="content">
+        <div class="ui hidden negative message" id="regerror">
+          <i class="close icon"></i>
+          <div class="header">
+            <span id="regerrormessage"></span>
+          </div>
+        </div>
+        <form class="ui form">
+          <div class="field">
+            <label>Username</label>
+            <input type="text" id="usernamereg" name="usernamereg" title="Max 15 characters/valid characters A-Z a-z 0-9 and _" pattern="[A-Za-z0-9_]{5,15}"/>
+          </div>
+          <div class="field">
+            <label>Email</label>
+            <input type="email" id="emailreg" name="emailreg"/>
+          </div>
+          <div class="field">
+            <label>Password</label>
+            <input type="password" id="passwdreg" name="passwdreg"/>
+          </div>
+          <div class="field">
+            <label>Password confirmation</label>
+            <input type="password" id="passwdregconf" name="passwdregconf"/>
+          </div>
+          <button class="ui fluid button" type="submit">Register</button>
+        </form>
+      </div>
+    </div>
+    </th:block>
+
+  </body>
+</html>
